+++ selenium-test-lens-examples/src/test/java/io/github/testlens/examples/OverlayPolicyExampleTest.java
@@ -2,13 +2,6 @@ package io.github.testlens.examples;
 
 import io.github.testlens.JsOverlayDebug;
 import io.github.testlens.OverlayConfig;
-import io.github.testlens.api.ApiCallActions;
-import io.github.testlens.api.ApiOverlayPanel;
-import io.github.testlens.core.Guards;
-import io.github.testlens.core.OverlayLogger;
-import io.github.testlens.core.OverlayRootManager;
-import io.github.testlens.core.logging.UiTestLensLogger;
-import io.github.testlens.selenium.SeleniumOverlayFactory;
 import io.github.testlens.selenium.overlay.OverlayAction;
 import io.github.testlens.selenium.overlay.OverlayHandler;
 import io.github.testlens.selenium.overlay.OverlayPolicy;
@@ -49,19 +42,7 @@ class OverlayPolicyExampleTest {
                 .build();
 
         OverlayConfig config = OverlayConfig.builder().build();
-        UiTestLensLogger eventLogger = UiTestLensLogger.builder().build();
-        OverlayLogger overlayLogger = OverlayLogger.from(eventLogger);
-        OverlayRootManager rootManager = SeleniumOverlayFactory.overlayRoot(driver, config);
-        ApiOverlayPanel apiPanel = SeleniumOverlayFactory.apiOverlayPanel(driver, rootManager, config);
-
-        JsOverlayDebug overlay = new JsOverlayDebug(
-                driver,
-                config,
-                apiPanel,
-                new ApiCallActions(apiPanel),
-                new Guards(driver, overlayLogger),
-                overlayLogger
-        );
+        JsOverlayDebug overlay = new JsOverlayDebug(driver, config);
         // In application code, call:
         // overlay.setOverlayPolicy(policy);
         JsOverlayDebug.class.getMethod("setOverlayPolicy", OverlayPolicy.class)