+++ selenium-test-lens-selenium/src/main/java/io/github/testlens/core/ScriptExecutor.java
@@ -1,5 +1,5 @@
 package io.github.testlens.core;
 
-public class ScriptExecutor {
+final class ScriptExecutor {
 }