+++ selenium-test-lens-selenium/src/main/java/io/github/testlens/selenium/locator/UiLocatorStatus.java
@@ -1,6 +1,6 @@
 package io.github.testlens.selenium.locator;
 
-public enum UiLocatorStatus {
+enum UiLocatorStatus {
     PASSED,
     FAILED,
     RETRYING